[Http] Refactor UriFactory by adding PsrUriFactory.
